it works! (almost)
