+++ src/details/r3d_drawcall.c
@@ -36,9 +36,9 @@
 /* === Internal functions === */
 
 // Functions applying OpenGL states defined by the material but unrelated to shaders
-static void r3d_drawcall_apply_depth_mode(R3D_DepthMode mode);
 static void r3d_drawcall_apply_cull_mode(R3D_CullMode mode);
 static void r3d_drawcall_apply_blend_mode(R3D_BlendMode mode);
+static void r3d_drawcall_apply_depth_mode(R3D_DepthMode mode);
 static void r3d_drawcall_apply_shadow_cast_mode(R3D_ShadowCastMode castMode, R3D_CullMode cullMode);
 
 // This function supports instanced rendering when necessary
@@ -68,9 +68,9 @@ bool r3d_drawcall_geometry_is_visible(const r3d_drawcall_t* call)
 {
     if (call->geometryType == R3D_DRAWCALL_GEOMETRY_MODEL) {
         if (r3d_matrix_is_identity(&call->transform)) {
-            return r3d_frustum_is_aabb_in(&R3D.state.frustum.shape, &call->geometry.model.mesh->aabb);
+            return r3d_frustum_is_aabb_in(&R3D.state.frustum.shape, &call->geometry.model.mesh.aabb);
         }
-        return r3d_frustum_is_obb_in(&R3D.state.frustum.shape, &call->geometry.model.mesh->aabb, &call->transform);
+        return r3d_frustum_is_obb_in(&R3D.state.frustum.shape, &call->geometry.model.mesh.aabb, &call->transform);
     }
 
     if (call->geometryType == R3D_DRAWCALL_GEOMETRY_SPRITE) {
@@ -91,35 +91,6 @@ bool r3d_drawcall_instanced_geometry_is_visible(const r3d_drawcall_t* call)
     }
 
     return r3d_frustum_is_obb_in(&R3D.state.frustum.shape, &call->instanced.allAabb, &call->transform);
-}
-
-void r3d_drawcall_update_model_animation(const r3d_drawcall_t* call)
-{
-    if (call->geometryType != R3D_DRAWCALL_GEOMETRY_MODEL || call->geometry.model.anim == NULL) {
-        return;
-    }
-
-    // skip animation update if custom is being used
-    if (call->geometry.model.boneOverride != NULL) {
-        return;
-    }
-
-    if (call->geometry.model.mesh->boneMatrices == NULL) {
-        // Only meshes belonging to a model with bones have a boneMatrices cache
-        TraceLog(LOG_WARNING, "Attempting to play animation on mesh without bone matrix cache");
-    }
-
-    int frame = call->geometry.model.frame;
-    if (frame >= call->geometry.model.anim->frameCount) {
-        frame = frame % call->geometry.model.anim->frameCount;
-    }
-
-    r3d_matrix_multiply_batch(
-        call->geometry.model.mesh->boneMatrices,
-        call->geometry.model.boneOffsets,
-        call->geometry.model.anim->frameGlobalPoses[frame],
-        call->geometry.model.anim->boneCount
-    );
 }
 
 void r3d_drawcall_raster_depth(const r3d_drawcall_t* call, bool forward, bool shadow, const Matrix* matVP)
@@ -147,7 +118,7 @@ void r3d_drawcall_raster_depth(const r3d_drawcall_t* call, bool forward, bool sh
     case R3D_DRAWCALL_GEOMETRY_MODEL:
         {
             // Send bone matrices and animation related data
-            if (call->geometry.model.anim != NULL && call->geometry.model.boneOffsets != NULL) {
+            if (call->geometry.model.player != NULL || R3D_IsSkeletonValid(&call->geometry.model.skeleton)) {
                 r3d_shader_set_int(raster.depth, uSkinning, true);
                 r3d_drawcall_upload_matrices(call);
             }
@@ -228,7 +199,7 @@ void r3d_drawcall_raster_depth_cube(const r3d_drawcall_t* call, bool forward, bo
     case R3D_DRAWCALL_GEOMETRY_MODEL:
         {
             // Send bone matrices and animation related data
-            if (call->geometry.model.anim != NULL && call->geometry.model.boneOffsets != NULL) {
+            if (call->geometry.model.player != NULL || R3D_IsSkeletonValid(&call->geometry.model.skeleton)) {
                 r3d_shader_set_int(raster.depthCube, uSkinning, true);
                 r3d_drawcall_upload_matrices(call);
             }
@@ -409,7 +380,7 @@ void r3d_drawcall_raster_geometry(const r3d_drawcall_t* call, const Matrix* matV
     case R3D_DRAWCALL_GEOMETRY_MODEL:
         {
             // Send bone matrices and animation related data
-            if (call->geometry.model.anim != NULL && call->geometry.model.boneOffsets != NULL) {
+            if (call->geometry.model.player != NULL || R3D_IsSkeletonValid(&call->geometry.model.skeleton)) {
                 r3d_shader_set_int(raster.geometry, uSkinning, true);
                 r3d_drawcall_upload_matrices(call);
             }
@@ -502,7 +473,7 @@ void r3d_drawcall_raster_forward(const r3d_drawcall_t* call, const Matrix* matVP
     case R3D_DRAWCALL_GEOMETRY_MODEL:
         {
             // Send bone matrices and animation related data
-            if (call->geometry.model.anim != NULL && call->geometry.model.boneOffsets != NULL) {
+            if (call->geometry.model.player != NULL || R3D_IsSkeletonValid(&call->geometry.model.skeleton)) {
                 r3d_shader_set_int(raster.forward, uSkinning, true);
                 r3d_drawcall_upload_matrices(call);
             }
@@ -553,29 +524,9 @@ void r3d_drawcall_raster_forward(const r3d_drawcall_t* call, const Matrix* matVP
 
 /* === Internal functions === */
 
-void r3d_drawcall_apply_depth_mode(R3D_DepthMode mode)
-{
-    switch (mode)
-    {
-    case R3D_DEPTH_DISABLED:
-        glDisable(GL_DEPTH_TEST);
-        break;
-    case R3D_DEPTH_READ_ONLY:
-        glEnable(GL_DEPTH_TEST);
-        glDepthMask(GL_FALSE);
-        break;
-    default:
-    case R3D_DEPTH_READ_WRITE:
-        glEnable(GL_DEPTH_TEST);
-        glDepthMask(GL_TRUE);
-        break;
-    }
-}
-
 void r3d_drawcall_apply_cull_mode(R3D_CullMode mode)
 {
-    switch (mode)
-    {
+    switch (mode) {
     case R3D_CULL_NONE:
         glDisable(GL_CULL_FACE);
         break;
@@ -592,8 +543,7 @@ void r3d_drawcall_apply_cull_mode(R3D_CullMode mode)
 
 void r3d_drawcall_apply_blend_mode(R3D_BlendMode mode)
 {
-    switch (mode)
-    {
+    switch (mode) {
     case R3D_BLEND_OPAQUE:
         glDisable(GL_BLEND);
         break;
@@ -618,10 +568,27 @@ void r3d_drawcall_apply_blend_mode(R3D_BlendMode mode)
     }
 }
 
+void r3d_drawcall_apply_depth_mode(R3D_DepthMode mode)
+{
+    switch (mode) {
+    case R3D_DEPTH_DISABLED:
+        glDisable(GL_DEPTH_TEST);
+        break;
+    case R3D_DEPTH_READ_ONLY:
+        glEnable(GL_DEPTH_TEST);
+        glDepthMask(GL_FALSE);
+        break;
+    default:
+    case R3D_DEPTH_READ_WRITE:
+        glEnable(GL_DEPTH_TEST);
+        glDepthMask(GL_TRUE);
+        break;
+    }
+}
+
 static void r3d_drawcall_apply_shadow_cast_mode(R3D_ShadowCastMode castMode, R3D_CullMode cullMode)
 {
-    switch (castMode)
-    {
+    switch (castMode) {
     case R3D_SHADOW_CAST_ON_AUTO:
         r3d_drawcall_apply_cull_mode(cullMode);
         break;
@@ -702,13 +669,13 @@ static void r3d_drawcall_unbind_geometry_mesh(void)
 void r3d_drawcall(const r3d_drawcall_t* call)
 {
     if (call->geometryType == R3D_DRAWCALL_GEOMETRY_MODEL) {
-        r3d_drawcall_bind_geometry_mesh(call->geometry.model.mesh);
-        r3d_drawcall_apply_depth_mode(call->geometry.model.mesh->depthMode);
-        if (call->geometry.model.mesh->indices == NULL) {
-            glDrawArrays(GL_TRIANGLES, 0, call->geometry.model.mesh->vertexCount);
+        r3d_drawcall_bind_geometry_mesh(&call->geometry.model.mesh);
+        r3d_drawcall_apply_depth_mode(call->material.depthMode);
+        if (call->geometry.model.mesh.ebo == 0) {
+            glDrawArrays(GL_TRIANGLES, 0, call->geometry.model.mesh.vertexCount);
         }
         else {
-            glDrawElements(GL_TRIANGLES, call->geometry.model.mesh->indexCount, GL_UNSIGNED_INT, NULL);
+            glDrawElements(GL_TRIANGLES, call->geometry.model.mesh.indexCount, GL_UNSIGNED_INT, NULL);
         }
         r3d_drawcall_unbind_geometry_mesh();
     }
@@ -724,8 +691,8 @@ void r3d_drawcall_instanced(const r3d_drawcall_t* call, int locInstanceModel, in
     // Bind the geometry
     switch (call->geometryType) {
     case R3D_DRAWCALL_GEOMETRY_MODEL:
-        r3d_drawcall_bind_geometry_mesh(call->geometry.model.mesh);
-        r3d_drawcall_apply_depth_mode(call->geometry.model.mesh->depthMode);
+        r3d_drawcall_bind_geometry_mesh(&call->geometry.model.mesh);
+        r3d_drawcall_apply_depth_mode(call->material.depthMode);
         break;
     case R3D_DRAWCALL_GEOMETRY_SPRITE:
         r3d_primitive_bind(&R3D.primitive.quad);
@@ -760,11 +727,11 @@ void r3d_drawcall_instanced(const r3d_drawcall_t* call, int locInstanceModel, in
     // Draw the geometry
     switch (call->geometryType) {
     case R3D_DRAWCALL_GEOMETRY_MODEL:
-        if (call->geometry.model.mesh->indices == NULL) {
-            glDrawArraysInstanced(GL_TRIANGLES, 0, call->geometry.model.mesh->vertexCount, (int)call->instanced.count);
+        if (call->geometry.model.mesh.ebo == 0) {
+            glDrawArraysInstanced(GL_TRIANGLES, 0, call->geometry.model.mesh.vertexCount, (int)call->instanced.count);
         }
         else {
-            glDrawElementsInstanced(GL_TRIANGLES, call->geometry.model.mesh->indexCount, GL_UNSIGNED_INT, NULL, (int)call->instanced.count);
+            glDrawElementsInstanced(GL_TRIANGLES, call->geometry.model.mesh.indexCount, GL_UNSIGNED_INT, NULL, (int)call->instanced.count);
         }
         break;
     case R3D_DRAWCALL_GEOMETRY_SPRITE:
@@ -802,9 +769,9 @@ static float r3d_drawcall_calculate_center_distance_to_camera(const r3d_drawcall
 {
     Vector3 center = { 0 };
     if (drawCall->geometryType == R3D_DRAWCALL_GEOMETRY_MODEL) {
-        center.x = (drawCall->geometry.model.mesh->aabb.min.x + drawCall->geometry.model.mesh->aabb.max.x) * 0.5f;
-        center.y = (drawCall->geometry.model.mesh->aabb.min.y + drawCall->geometry.model.mesh->aabb.max.y) * 0.5f;
-        center.z = (drawCall->geometry.model.mesh->aabb.min.z + drawCall->geometry.model.mesh->aabb.max.z) * 0.5f;
+        center.x = (drawCall->geometry.model.mesh.aabb.min.x + drawCall->geometry.model.mesh.aabb.max.x) * 0.5f;
+        center.y = (drawCall->geometry.model.mesh.aabb.min.y + drawCall->geometry.model.mesh.aabb.max.y) * 0.5f;
+        center.z = (drawCall->geometry.model.mesh.aabb.min.z + drawCall->geometry.model.mesh.aabb.max.z) * 0.5f;
     }
     else if (drawCall->geometryType == R3D_DRAWCALL_GEOMETRY_SPRITE) {
         center.x = drawCall->transform.m12;
@@ -825,14 +792,14 @@ static float r3d_drawcall_calculate_max_distance_to_camera(const r3d_drawcall_t*
     }
 
     Vector3 corners[8] = {
-        {drawCall->geometry.model.mesh->aabb.min.x, drawCall->geometry.model.mesh->aabb.min.y, drawCall->geometry.model.mesh->aabb.min.z},
-        {drawCall->geometry.model.mesh->aabb.max.x, drawCall->geometry.model.mesh->aabb.min.y, drawCall->geometry.model.mesh->aabb.min.z},
-        {drawCall->geometry.model.mesh->aabb.min.x, drawCall->geometry.model.mesh->aabb.max.y, drawCall->geometry.model.mesh->aabb.min.z},
-        {drawCall->geometry.model.mesh->aabb.max.x, drawCall->geometry.model.mesh->aabb.max.y, drawCall->geometry.model.mesh->aabb.min.z},
-        {drawCall->geometry.model.mesh->aabb.min.x, drawCall->geometry.model.mesh->aabb.min.y, drawCall->geometry.model.mesh->aabb.max.z},
-        {drawCall->geometry.model.mesh->aabb.max.x, drawCall->geometry.model.mesh->aabb.min.y, drawCall->geometry.model.mesh->aabb.max.z},
-        {drawCall->geometry.model.mesh->aabb.min.x, drawCall->geometry.model.mesh->aabb.max.y, drawCall->geometry.model.mesh->aabb.max.z},
-        {drawCall->geometry.model.mesh->aabb.max.x, drawCall->geometry.model.mesh->aabb.max.y, drawCall->geometry.model.mesh->aabb.max.z}
+        {drawCall->geometry.model.mesh.aabb.min.x, drawCall->geometry.model.mesh.aabb.min.y, drawCall->geometry.model.mesh.aabb.min.z},
+        {drawCall->geometry.model.mesh.aabb.max.x, drawCall->geometry.model.mesh.aabb.min.y, drawCall->geometry.model.mesh.aabb.min.z},
+        {drawCall->geometry.model.mesh.aabb.min.x, drawCall->geometry.model.mesh.aabb.max.y, drawCall->geometry.model.mesh.aabb.min.z},
+        {drawCall->geometry.model.mesh.aabb.max.x, drawCall->geometry.model.mesh.aabb.max.y, drawCall->geometry.model.mesh.aabb.min.z},
+        {drawCall->geometry.model.mesh.aabb.min.x, drawCall->geometry.model.mesh.aabb.min.y, drawCall->geometry.model.mesh.aabb.max.z},
+        {drawCall->geometry.model.mesh.aabb.max.x, drawCall->geometry.model.mesh.aabb.min.y, drawCall->geometry.model.mesh.aabb.max.z},
+        {drawCall->geometry.model.mesh.aabb.min.x, drawCall->geometry.model.mesh.aabb.max.y, drawCall->geometry.model.mesh.aabb.max.z},
+        {drawCall->geometry.model.mesh.aabb.max.x, drawCall->geometry.model.mesh.aabb.max.y, drawCall->geometry.model.mesh.aabb.max.z}
     };
 
     float maxDistSq = 0.0f;
@@ -879,17 +846,17 @@ static void r3d_drawcall_upload_matrices(const r3d_drawcall_t* call)
 
     const int bindingSlot = 0;
 
-    if (call->geometry.model.boneOverride == NULL) {
+    if (call->geometry.model.player != NULL) {
         r3d_storage_bind_and_upload_matrices(
-            call->geometry.model.mesh->boneMatrices,
-            call->geometry.model.mesh->boneCount,
+            call->geometry.model.player->currentPose,
+            call->geometry.model.player->skeleton.boneCount,
             bindingSlot
         );
     }
     else {
         r3d_storage_bind_and_upload_matrices(
-            call->geometry.model.boneOverride,
-            call->geometry.model.anim->boneCount,
+            call->geometry.model.skeleton.bindPose,
+            call->geometry.model.skeleton.boneCount,
             bindingSlot
         );
     }
