+++ src/r3d_curves.c
@@ -1,9 +1,20 @@
-#include "r3d.h"
+/* r3d_curves.h -- R3D Curves Module.
+ *
+ * Copyright (c) 2025 Le Juez Victor
+ *
+ * This software is provided 'as-is', without any express or implied warranty.
+ * For conditions of distribution and use, see accompanying LICENSE file.
+ */
+
+#include <r3d/r3d_curves.h>
+#include <raymath.h>
+#include <stdlib.h>
 
 #include "./details/containers/r3d_array.h"
 
-#include <raymath.h>
-#include <stdlib.h>
+// ========================================
+// PUBLIC API
+// ========================================
 
 R3D_InterpolationCurve R3D_LoadInterpolationCurve(int capacity)
 {
