+++ src/r3d_environment.c
@@ -1,26 +1,19 @@
-/*
+/* r3d_environment.h -- R3D Environment Module.
+ *
  * Copyright (c) 2025 Le Juez Victor
  *
- * This software is provided "as-is", without any express or implied warranty. In no event
- * will the authors be held liable for any damages arising from the use of this software.
- *
- * Permission is granted to anyone to use this software for any purpose, including commercial
- * applications, and to alter it and redistribute it freely, subject to the following restrictions:
- *
- *   1. The origin of this software must not be misrepresented; you must not claim that you
- *   wrote the original software. If you use this software in a product, an acknowledgment
- *   in the product documentation would be appreciated but is not required.
- *
- *   2. Altered source versions must be plainly marked as such, and must not be misrepresented
- *   as being the original software.
- *
- *   3. This notice may not be removed or altered from any source distribution.
+ * This software is provided 'as-is', without any express or implied warranty.
+ * For conditions of distribution and use, see accompanying LICENSE file.
  */
 
-#include "r3d.h"
+#include <r3d/r3d_environment.h>
+#include <raymath.h>
 
 #include "./r3d_state.h"
-#include <raymath.h>
+
+// ========================================
+// PUBLIC API
+// ========================================
 
 void R3D_SetBackgroundColor(Color color)
 {
