+++ src/r3d_lighting.c
@@ -1,32 +1,20 @@
-/*
+/* r3d_lighting.h -- R3D Lighting Module.
+ *
  * Copyright (c) 2025 Le Juez Victor
  *
- * This software is provided "as-is", without any express or implied warranty. In no event
- * will the authors be held liable for any damages arising from the use of this software.
- *
- * Permission is granted to anyone to use this software for any purpose, including commercial
- * applications, and to alter it and redistribute it freely, subject to the following restrictions:
- *
- *   1. The origin of this software must not be misrepresented; you must not claim that you
- *   wrote the original software. If you use this software in a product, an acknowledgment
- *   in the product documentation would be appreciated but is not required.
- *
- *   2. Altered source versions must be plainly marked as such, and must not be misrepresented
- *   as being the original software.
- *
- *   3. This notice may not be removed or altered from any source distribution.
+ * This software is provided 'as-is', without any express or implied warranty.
+ * For conditions of distribution and use, see accompanying LICENSE file.
  */
 
-#include "r3d.h"
+#include <r3d/r3d_lighting.h>
+#include <raymath.h>
 
 #include "./details/r3d_light.h"
 #include "./r3d_state.h"
 
-#include <raylib.h>
-#include <raymath.h>
-
-
-/* === Helper macros === */
+// ========================================
+// HELPER MACROS
+// ========================================
 
 #define r3d_get_and_check_light(var_name, id, ...)                  \
     r3d_light_t* var_name;                                          \
@@ -38,8 +26,9 @@
     }                                                               \
 }
 
-
-/* === Public functions === */
+// ========================================
+// PUBLIC API
+// ========================================
 
 R3D_Light R3D_CreateLight(R3D_LightType type)
 {
