+++ src/r3d_state.h
@@ -21,6 +21,7 @@
 #define R3D_STATE_H
 
 #include <r3d/r3d_environment.h>
+#include <r3d/r3d_material.h>
 #include <r3d/r3d_core.h>
 #include <r3d/r3d_mesh.h>
 #include <glad.h>
@@ -340,7 +341,8 @@ extern struct R3D_State {
     // Misc data
     struct {
         Matrix matCubeViews[6];
-        R3D_Mesh meshDecalBounds;     //< Unit cube used for decal projection
+        // REVIEW: We could use internal primitive rendering (?)
+        R3D_Mesh meshDecalBounds;   //< Unit cube used for decal projection
     } misc;
 
 } R3D;
